+++ src/imio/smartweb/core/webcomponents/src/components/Events/ContactCard/ContactCard.jsx
@@ -19,12 +19,12 @@ const ContactCard = ({ contactItem }) => {
     const topics = contactItem.topics ? contactItem.topics : "";
     return (
         <div className="r-list-item">
-            <div
-                className="r-item-img"
+            <div    
+                className={contactItem.image?"r-item-img":"r-item-img r-item-img-placeholder"}
                 style={{
                     backgroundImage: contactItem.image
                         ? "url(" + contactItem.image.scales.preview.download + ")"
-                        : "url(" + imgPlaceholder + ")",
+                        : "",
                 }}
             />