+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.10 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Remove forced placeholder for image in react pages
+  [thomlamb]
 
 
 1.0.9 (2022-05-02)